rewrite JSONP test file to run against worker
