+++ package.json
@@ -42,7 +42,7 @@
     "start": "node server/run.js",
     "lint": "eslint .",
     "client_test": "phantomjs ./node_modules/mocha-phantomjs-core/mocha-phantomjs-core.js test.html spec '{\"useColors\":true}'",
-    "server_test": "mocha test/server/worker_test.js",
+    "server_test": "mocha test/server/",
     "test": "npm run lint && npm run server_test && npm run client_test"
   }
 }
