+++ test/server/worker_test.js
@@ -25,22 +25,6 @@ describe('CloudFlare Worker', () => {
     expect(res.headers.get('x-foo')).to.eql('bar');
   });
 
-  it('supports JSONP', async () => {
-    const destHost = 'http://localhost:8001';
-    const json = { message: 'test' };
-    nock(destHost)
-      .get('/')
-      .reply(200, json);
-
-    const req = new Request(`http://jsonp.test/?url=${destHost}&callback=foo`);
-    const res = await handleRequest(req);
-
-    expect(res.status).to.be(200);
-    expect(res.headers.get('content-type')).to.eql('text/javascript');
-    const data = await res.text();
-    expect(data).to.eql('foo({"message":"test"});');
-  });
-
   it('renders the homepage', async () => {
     const req = new Request(`http://jsonp.test/`);
     const res = await handleRequest(req);
